+++ NWN.Anvil/src/main/API/Events/Game/DoorEvents/DoorEvents.OnDeath.cs
@@ -1,0 +1,44 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for door objects.
+  /// </summary>
+  public static partial class DoorEvents
+  {
+    [GameEvent(EventScriptType.DoorOnDeath)]
+    public sealed class OnDeath : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwDoor"/> that was destroy.
+      /// </summary>
+      public NwDoor Door { get; } = NWScript.OBJECT_SELF.ToNwObject<NwDoor>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that killed the <see cref="NwDoor"/>.
+      /// </summary>
+      public NwCreature Killer { get; } = NWScript.GetLastKiller().ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Door;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwDoor
+  {
+    /// <inheritdoc cref="DoorEvents.OnDeath"/>
+    public event Action<DoorEvents.OnDeath> OnDeath
+    {
+      add => EventService.Subscribe<DoorEvents.OnDeath, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<DoorEvents.OnDeath, GameEventFactory>(this, value);
+    }
+  }
+}
