+++ NWN.Anvil/src/main/API/Events/Game/DoorEvents/DoorEvents.OnDialogue.cs
@@ -1,0 +1,39 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for door objects.
+  /// </summary>
+  public static partial class DoorEvents
+  {
+    [GameEvent(EventScriptType.DoorOnDialogue)]
+    public sealed class OnDialogue : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwDoor"/> that is in dialog.
+      /// </summary>
+      public NwDoor Door { get; } = NWScript.OBJECT_SELF.ToNwObject<NwDoor>();
+
+      NwObject IEvent.Context
+      {
+        get => Door;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwDoor
+  {
+    /// <inheritdoc cref="DoorEvents.OnDialogue"/>
+    public event Action<DoorEvents.OnDialogue> OnDialogue
+    {
+      add => EventService.Subscribe<DoorEvents.OnDialogue, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<DoorEvents.OnDialogue, GameEventFactory>(this, value);
+    }
+  }
+}
