+++ NWN.Anvil/src/main/API/Events/Game/DoorEvents/DoorEvents.OnDisarm.cs
@@ -1,0 +1,39 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for door objects.
+  /// </summary>
+  public static partial class DoorEvents
+  {
+    [GameEvent(EventScriptType.DoorOnDisarm)]
+    public sealed class OnDisarm : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwDoor"/> that was disarmed.
+      /// </summary>
+      public NwDoor Door { get; } = NWScript.OBJECT_SELF.ToNwObject<NwDoor>();
+
+      NwObject IEvent.Context
+      {
+        get => Door;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwDoor
+  {
+    /// <inheritdoc cref="DoorEvents.OnDisarm"/>
+    public event Action<DoorEvents.OnDisarm> OnDisarm
+    {
+      add => EventService.Subscribe<DoorEvents.OnDisarm, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<DoorEvents.OnDisarm, GameEventFactory>(this, value);
+    }
+  }
+}
