+++ NWN.Anvil/src/main/API/Events/Game/DoorEvents/DoorEvents.OnHeartbeat.cs
@@ -1,0 +1,39 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for door objects.
+  /// </summary>
+  public static partial class DoorEvents
+  {
+    [GameEvent(EventScriptType.DoorOnHeartbeat)]
+    public sealed class OnHeartbeat : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwDoor"/> that had a heartbeat.
+      /// </summary>
+      public NwDoor Door { get; } = NWScript.OBJECT_SELF.ToNwObject<NwDoor>();
+
+      NwObject IEvent.Context
+      {
+        get => Door;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwDoor
+  {
+    /// <inheritdoc cref="DoorEvents.OnHeartbeat"/>
+    public event Action<DoorEvents.OnHeartbeat> OnHeartbeat
+    {
+      add => EventService.Subscribe<DoorEvents.OnHeartbeat, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<DoorEvents.OnHeartbeat, GameEventFactory>(this, value);
+    }
+  }
+}
