+++ NWN.Anvil/src/main/API/Events/Game/DoorEvents/DoorEvents.OnUnlock.cs
@@ -1,0 +1,44 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for door objects.
+  /// </summary>
+  public static partial class DoorEvents
+  {
+    [GameEvent(EventScriptType.DoorOnUnlock)]
+    public sealed class OnUnlock : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwDoor"/> that was unlocked.
+      /// </summary>
+      public NwDoor Door { get; } = NWScript.OBJECT_SELF.ToNwObject<NwDoor>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that unlocked <see cref="NwDoor"/>.
+      /// </summary>
+      public NwCreature UnlockedBy { get; } = NWScript.GetLastUnlocked().ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Door;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwDoor
+  {
+    /// <inheritdoc cref="DoorEvents.OnUnlock"/>
+    public event Action<DoorEvents.OnUnlock> OnUnlock
+    {
+      add => EventService.Subscribe<DoorEvents.OnUnlock, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<DoorEvents.OnUnlock, GameEventFactory>(this, value);
+    }
+  }
+}
